+++ src/models/DoanhThuThang.js
@@ -25,7 +25,7 @@ module.exports = (sequelize, DataTypes) => {
                 type: DataTypes.INTEGER,
             },
             SoChuyenBay: DataTypes.INTEGER,
-            DoanhThu: DataTypes.INTEGER,
+            DoanhThu: DataTypes.BIGINT,
         },
         {
             sequelize,
